requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
asgiref==3.9.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.6
django-fsm==3.0.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
pillow==12.0.0
propcache==0.4.1
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
schedule==1.2.2
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
yarl==1.22.0
